Removed requirements folder for now
Will bring it back when MetalBase 1.0 is finished. Initially I was
planning to have an SRS, but I decided that the Prototype would get the
job done better. I may write one when the first version of MetalBase is
finished so that I remember the requirements if I decide to come back
later and expand/modify MetalBase